requirements.txt
streamlit
openai
python-dotenv
duckduckgo-search
wolframalpha
wikipedia